feat: move crate location and some fix
